requirements.txt
streamlit==0.80.0
Pillow==8.2.0
matplotlib==3.4.1
numpy==1.19.5
scipy==1.6.2
numpy==1.19.5
mtcnn==0.1.0
tensorflow==2.4.1
keras-vggface==0.6
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
opencv-python==4.5.1.48